some data types fixed

+++ src/main/kotlin/com/chrzanekk/simulatorpandemic/payload/NewSimulationRequest.kt
@@ -1,11 +1,13 @@
 package com.chrzanekk.simulatorpandemic.payload
 
+import java.math.BigDecimal
+
 data class NewSimulationRequest(
     val simulationName: String,
-    val populationSize: Int,
-    val initialNoOfInfectedPpl: Int,
-    val infectionRate: Float,
-    val mortalityRate: Float,
+    val populationSize: BigDecimal,
+    val initialNoOfInfectedPpl: BigDecimal,
+    val infectionRate: BigDecimal,
+    val mortalityRate: BigDecimal,
     val infectedToCureDuration: Int,
     val infectedToDeathDuration: Int,
     val simulationDays: Int
