+++ src/main/kotlin/com/chrzanekk/simulatorpandemic/payload/NewSimulationResponse.kt
@@ -1,13 +1,14 @@
 package com.chrzanekk.simulatorpandemic.payload
 
 import com.chrzanekk.simulatorpandemic.domain.SimulatedPopulation
+import java.math.BigDecimal
 
 data class NewSimulationResponse(
     val simulationName: String,
-    val populationSize: Int,
-    val initialNoOfInfectedPpl: Int,
-    val infectionRate: Float,
-    val mortalityRate: Float,
+    val populationSize: BigDecimal,
+    val initialNoOfInfectedPpl: BigDecimal,
+    val infectionRate: BigDecimal,
+    val mortalityRate: BigDecimal,
     val infectedToCureDuration: Int,
     val infectedToDeathDuration: Int,
     val simulationDays: Int,
